# server/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.RackRequest:
    properties:
      description:
        example: Left side
        type: string
      max_capacity:
        example: 12
        type: integer
      name:
        example: Top shelf
        type: string
    type: object
...
  models.Rack:
    properties:
      created_at:
        type: string
      description:
        type: string
      id_rack:
        type: integer
      id_user:
        type: string
      max_capacity:
        type: integer
      name:
        type: string
      spool_count:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Filament inventory management for 3D printing enthusiasts.
  title: Filaventory API
  version: "1.0"
paths:
...
  /api/racks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Rack'
            type: array
      security:
      - BearerAuth: []
      summary: List racks
      tags:
      - racks
    post:
      consumes:
      - application/json
      parameters:
      - description: Rack data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RackRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Rack'
        "400":
          description: name is required
          schema:
            type: string
        "409":
          description: rack name already exists
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a rack
      tags:
      - racks
  /api/racks/{id}:
    delete:
      parameters:
      - description: Rack ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a rack
      tags:
      - racks
    get:
      parameters:
      - description: Rack ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rack'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a rack
      tags:
      - racks
    put:
      consumes:
      - application/json
      parameters:
      - description: Rack ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rack data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rack'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a rack
      tags:
      - racks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" in the Authorization header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"